Separate inputs from parent
