+++ pyrene/src/examples/index.ts
@@ -13,7 +13,7 @@ import { default as Checkbox } from '../components/Checkbox/Checkbox.examples';
 import { default as CheckboxPopover } from '../components/CheckboxPopover/CheckboxPopover.examples';
 import { default as Collapsible } from '../components/Collapsible/Collapsible.examples';
 import { default as Container } from '../components/Container/Container.examples';
-import { default as DateTimeInput } from '../components/DateTimeInput/DateTimeInput.examples';
+import { default as DateTimePicker } from '../components/DateTimePicker/DateTimePicker.examples';
 import { default as DropdownButton } from '../components/DropdownButton/DropdownButton.examples';
 import { default as Filter } from '../components/Filter/Filter.examples';
 import { default as Icon } from '../components/Icon/Icon.examples';
@@ -64,7 +64,7 @@ const examples: { [key: string]: Example<any, any> } = {
   CheckboxPopover,
   Collapsible,
   Container,
-  DateTimeInput,
+  DateTimePicker,
   DropdownButton,
   Filter,
   Heading,
